Fix specs for SplitTimeslice. Improve specs for UpdateCourseWikiTimeslices. Move timeslice comparing methods to a new timeslice helper.
