+++ spec/services/split_timeslice_spec.rb
@@ -1,72 +1,7 @@
 # frozen_string_literal: true
 
-require 'yaml'
 require 'rails_helper'
 
-def compare_course_wiki_timeslices(course, path)
-  cw_actual = course.course_wiki_timeslices.map do |ts|
-    ts.attributes.slice(
-      'start', 'end',
-      'character_sum', 'references_count',
-      'revision_count', 'stats',
-      'last_mw_rev_datetime', 'needs_update'
-    )
-  end
-
-  cw_expected = YAML.load_file(Rails.root + path + 'expected_timeslices.yml')
-
-  cw_expected.each_value do |ts|
-    # Ensure dates are timezone and not strings
-    ts['start'] = ts['start'].to_time.in_time_zone('UTC')
-    ts['end'] = ts['end'].to_time.in_time_zone('UTC')
-    if ts['last_mw_rev_datetime']
-      ts['last_mw_rev_datetime'] =
-        ts['last_mw_rev_datetime'].to_time.in_time_zone('UTC')
-    end
-  end
-
-  expect(cw_actual).to match_array(cw_expected.values)
-end
-
-def compare_article_course_timeslices(course, path)
-  ac_actual = course.article_course_timeslices.map do |ts|
-    ts.attributes.slice(
-      'start', 'end', 'character_sum', 'references_count',
-      'revision_count', 'new_article', 'tracked', 'first_revision', 'user_ids'
-    )
-  end
-
-  ac_expected = YAML.load_file(Rails.root + path + 'expected_ac_timeslices.yml')
-
-  ac_expected.each_value do |ts|
-    # Ensure dates are timezone and not strings
-    ts['start'] = ts['start'].to_time.in_time_zone('UTC')
-    ts['end'] = ts['end'].to_time.in_time_zone('UTC')
-    ts['first_revision'] = ts['first_revision'].to_time.in_time_zone('UTC')
-    ts['user_ids'] = [user.id]
-  end
-
-  expect(ac_actual).to match_array(ac_expected.values)
-end
-
-def compare_course_user_wiki_timeslices(course, path)
-  cuw_actual = course.course_user_wiki_timeslices.map do |ts|
-    ts.attributes.slice(
-      'start', 'end', 'character_sum_ms', 'character_sum_us',
-      'character_sum_draft', 'references_count', 'revision_count'
-    )
-  end
-
-  cuw_expected = YAML.load_file(Rails.root + path + 'expected_cuw_timeslices.yml')
-
-  cuw_expected.each_value do |ts|
-    # Ensure dates are timezone and not strings
-    ts['start'] = ts['start'].to_time.in_time_zone('UTC')
-    ts['end'] = ts['end'].to_time.in_time_zone('UTC')
-  end
-
-  expect(cuw_actual).to match_array(cuw_expected.values)
-end
 describe SplitTimeslice do
   let(:start) { '2025-08-26 17:00:00'.to_datetime }
   let(:end_date) { '2025-08-27 17:00:00'.to_datetime }
@@ -83,45 +18,56 @@ describe SplitTimeslice do
     JoinCourse.new(course:, user:, role: 0)
   end
 
-  describe '#handle' do
-    context 'when revisions exceed threshold' do
-      let(:path) { 'spec/support/split_timeslices/over_threshold/' }
+  describe '#maybe_split' do
+    let(:rev1) { build(:revision_on_memory, scoped: true) }
+    let(:rev2) { build(:revision_on_memory, scoped: true) }
+    let(:rev3) { build(:revision_on_memory, scoped: true) }
+    let(:revisions) { { wiki => { revisions: [rev1, rev2, rev3] } } }
 
+    context 'when revisions exceed threshold' do
       before do
-        stub_const('SplitTimeslice::REVISION_THRESHOLD', 10)
+        stub_const('SplitTimeslice::REVISION_THRESHOLD', 2)
         TimesliceManager.new(course).create_timeslices_for_new_course_wiki_records([wiki])
       end
 
-      it 'splits the timeslice recursively' do
+      it 'splits the timeslice' do
         expect(course.course_wiki_timeslices.count).to eq(2)
-        VCR.use_cassette 'split_timeslices/over_threshold' do
-          dates = splitter.handle(wiki, course.start, course.end)
-          expect(dates).to all(be_a(Time))
-          expect(dates.size).to eq(10)
-        end
-        compare_course_wiki_timeslices(course, path)
-        compare_article_course_timeslices(course, path)
-        compare_course_user_wiki_timeslices(course, path)
+        split, dates = splitter.maybe_split(wiki, course.start, course.start + 1.day, revisions)
+        expect(split).to eq(true)
+        expect(dates[0]).to eq(course.start)
+        expect(dates[1]).to eq(course.start + 12.hours)
+        expect(dates[2]).to eq(course.start + 1.day)
+        expect(dates).to all(be_a(Time))
+        # timeslice was deleted
+        expect(CourseWikiTimeslice.where(course:, wiki:, start:,
+                                         end: start + 1.day)).to be_empty
+      end
+
+      it 'splits the timeslice even when odd seconds' do
+        expect(course.course_wiki_timeslices.count).to eq(2)
+        split, dates = splitter.maybe_split(wiki, course.start, course.start + 675.seconds,
+                                            revisions)
+        expect(split).to eq(true)
+        expect(dates[0]).to eq(course.start)
+        expect(dates[1]).to eq(course.start + 338.seconds)
+        expect(dates[2]).to eq(course.start + 675.seconds)
+        expect(dates).to all(be_a(Time))
       end
     end
 
     context 'when revisions do not exceed threshold' do
-      let(:path) { 'spec/support/split_timeslices/under_threshold/' }
-
       before do
         TimesliceManager.new(course).create_timeslices_for_new_course_wiki_records([wiki])
       end
 
       it 'does not split the timeslice' do
         expect(course.course_wiki_timeslices.count).to eq(2)
-        VCR.use_cassette 'split_timeslices/under_threshold' do
-          dates = splitter.handle(wiki, course.start, course.end)
-          expect(dates).to all(be_a(Time))
-          expect(dates.size).to eq(1)
-        end
-        compare_course_wiki_timeslices(course, path)
-        compare_article_course_timeslices(course, path)
-        compare_course_user_wiki_timeslices(course, path)
+        split, dates = splitter.maybe_split(wiki, course.start, course.start + 12.hours, revisions)
+        expect(split).to eq(false)
+        expect(dates).to be_empty
+        # timeslice was created
+        expect(CourseWikiTimeslice.where(course:, wiki:, start:,
+                                         end: start + 12.hours).size).to eq(1)
       end
     end
   end
